Move memory-based operations to cliquet.storage.memory
