+++ cliquet/storage/cloud_storage.py
@@ -5,8 +5,8 @@ import six
 from requests.exceptions import RequestException
 
 from cliquet import logger
-from cliquet.storage import (StorageBase, exceptions, get_unicity_rules,
-                             Filter, apply_sorting)
+from cliquet.storage import StorageBase, exceptions, Filter
+from cliquet.storage.memory import apply_sorting, get_unicity_rules
 from cliquet.utils import json, COMPARISON
